% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args_cli.R
\name{args_cli}
\alias{args_cli}
\title{Get arguments from the command line interface}
\usage{
args_cli()
}
\value{
A vector of command line arguments (as other programming languages would return them, unlike the crummy R default)
}
\description{
A simple wrapper around \code{\link[=commandArgs]{commandArgs()}} that sets \code{trailingOnly = TRUE}, reducing typing.
}
\examples{
args <- args_cli()

}
