% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_submit.R
\name{batch_submit}
\alias{batch_submit}
\title{Submit an existing slurm batch script}
\usage{
batch_submit(name)
}
\arguments{
\item{name}{Either the full batch job file path, or its base name (excluding the extension \code{.q}).
Must exist.}
}
\description{
Given an existing batch script, this script calls \code{sbatch} with parameters and gets job submitted.
Ideal to simplify big loops.
}
\examples{

# set desired parameters
commands <- 'time echo "Hello world!"'
name <- 'hola'
mem <- '1G'

# writes file `hola.q`
batch_writer(
    commands,
    name,
    mem
)

# this submits `hola.q`:
batch_submit( 'hola' )

# cleanup
batch_cleanup( name )

}
\seealso{
\code{\link[=batch_writer]{batch_writer()}} to write slurm batch scripts.
\code{\link[=batch_cleanup]{batch_cleanup()}} to remove slurm batch scripts.
}
