% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_writer.R
\name{batch_writer}
\alias{batch_writer}
\title{Write a slurm batch script file}
\usage{
batch_writer(
  commands,
  name,
  R = "4.0.0",
  mem = NA,
  time = NA,
  nodes = 1,
  threads = 1,
  array = NA,
  partition = "biostat",
  account = NA,
  email = NA
)
}
\arguments{
\item{commands}{A character vector of commands to run, one per line.}

\item{name}{Short name for for batch script file, job, and stdout/stderr output paths}

\item{R}{Specify R version, to load module (\code{NA} prevents loading R).}

\item{mem}{Memory limit.
\code{NA} (default) is no memory limit.
May be a string like \verb{1G}, whatever sbatch accepts.}

\item{time}{Time limit.
\code{NA} (default) is no time limit.
May be a string like \code{1:00:00}, whatever sbatch accepts.}

\item{nodes}{Number of nodes to use (default 1).}

\item{threads}{Number of tasks per node (default 1).}

\item{array}{Value for array jobs (default NA is not an array job; typical values are numeric ranges passed as strings, such as "1-50", "0-90", or discontinuous ranges such as "1,4,10-20").}

\item{partition}{The partition to submit to, by default \code{biostat}.
If \code{NA} then no partition is declared (default partition will be used, \code{common} in DCC?).}

\item{account}{Account to use.
\code{NA} (default) is default account for user.}

\item{email}{Address to receive emails about job completion.}
}
\description{
A wrapper around writing the script given the desired command, a short name for the job, and many optional parameters.
Defaults are for my current setup (latest R version on cluster).
}
\examples{

# set desired parameters
commands <- 'time echo "Hello world!"'
name <- 'hola'

# writes file `hola.q`
batch_writer(
    commands,
    name
)

# cleanup
batch_cleanup( name )

}
\seealso{
\code{\link[=batch_submit]{batch_submit()}} to submit slurm batch scripts.
\code{\link[=batch_cleanup]{batch_cleanup()}} to remove slurm batch scripts.
}
