% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_end.R
\name{fig_end}
\alias{fig_end}
\title{Ends a PDF figure}
\usage{
fig_end()
}
\description{
\code{fig_end} closes the PDF device created by \code{fig_start}.
It's simply a wrapper to \verb{\link[grDevices]\{dev.off\}} that returns invisibly and saves on typing.
}
\examples{
# fig_start('my_plot')
# plot(1:10)
# fig_end()

}
