% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_width.R
\name{fig_width}
\alias{fig_width}
\title{Figure widths requirements for several journals}
\usage{
fig_width(journal = c("plos", "pnas"), size = c("max", "min", "mid"))
}
\arguments{
\item{journal}{String specifying journal of interest}

\item{size}{Desired size range ("max" and "min" are always defined, "mid" is non-NA in some cases).}
}
\description{
Returns by default maximum width values in inches (~ 7-7.5) for figures according to various journal specifications.
It also optionally returns minimum (~ 2.6-3.5) and mid values (~ 4.3), for fractional column sizes, if defined.
}
\examples{
fig_width('plos')

}
\seealso{
\code{\link[=fig_height]{fig_height()}}

PLoS Genetics reference:
\url{https://journals.plos.org/plosgenetics/s/figures}

PNAS reference:
\url{https://www.pnas.org/authors/submitting-your-manuscript#manuscript-formatting-guidelines}

Journals that don't specify limits:

Genetics reference:
\url{https://www.genetics.org/content/prep-manuscript#figures}

Nature Genetics reference:
\url{https://www.nature.com/ng/for-authors/preparing-your-submission#formatting}
}
