% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_submit.R
\name{batch_submit}
\alias{batch_submit}
\title{Submit an existing batch job file}
\usage{
batch_submit(name, partition = "biostat")
}
\arguments{
\item{name}{Either the full batch job file path, or its base name (excluding the extension \code{.q}).
Must exist.}

\item{partition}{The partition to submit to, by default \code{biostat}.}
}
\description{
Given an existing sbatch job file, this script calls \code{sbatch} with parameters and gets job submitted.
Ideal to simplify big loops.
}
\examples{

# set desired parameters
commands <- 'time echo "Hello world!"'
name <- 'hola'
mem <- '1G'

# writes file `hola.q`
batch_writer(
    commands,
    name,
    mem
)

# this submits `hola.q` to `biostat` partition:
batch_submit( 'hola' )

# cleanup
file.remove( paste0( name, '.q' ) )

}
\seealso{
\verb{\link{batch_writer}} to write these batch files
}
