% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_start.R
\name{fig_start}
\alias{fig_start}
\title{Starts a PDF or PNG figure with better defaults}
\usage{
fig_start(
  name,
  width = 3,
  height = 3,
  bg = "white",
  mar_b = 3,
  mar_l = 3,
  mar_t = 0,
  mar_r = 0,
  mar = NULL,
  mar_pad = 0.2,
  png = FALSE,
  png_res = 600,
  verbose = TRUE
)
}
\arguments{
\item{name}{The output path, excluding extension (\code{.pdf} or \code{.png} is automatically added)}

\item{width}{Figure width, in inches (default 3).
If \code{png = TRUE}, this value gets multiplied by \code{png_res} to result in the final width in pixels.}

\item{height}{Figure height, in inches (default 3).
If \code{png = TRUE}, this value gets multiplied by \code{png_res} to result in the final height in pixels.}

\item{bg}{Figure background color (default "white")}

\item{mar_b}{Bottom margin size, in margin lines (default 3).
\code{mar_pad} gets added to this value to produce the final margins.}

\item{mar_l}{Left margin size, in margin lines (default 3).
\code{mar_pad} gets added to this value to produce the final margins.}

\item{mar_t}{Top margin size, in margin lines (default 0).
\code{mar_pad} gets added to this value to produce the final margins.}

\item{mar_r}{Right margin size, in margin lines (default 0).
\code{mar_pad} gets added to this value to produce the final margins.}

\item{mar}{A vector with four elements (in the format \code{\link[graphics:par]{graphics::par()}} expects, as an alternative specification.
\code{mar_pad} gets added to this vector to produce the final margins.
The individual margins above are used if \code{mar = NULL} (default).}

\item{mar_pad}{Margin padding applied to all margins.}

\item{png}{If \code{TRUE}, output is a PNG file (default \code{FALSE} results in a PDF file).}

\item{png_res}{Resolution of PNG image in ppi.
In addition to setting the resolution internally in the file, this scales the \code{width}and \code{height} values (default in inches) to the final pixel values of the output.
Ignored if \code{png = FALSE}.}

\item{verbose}{If \code{TRUE} (default), prints a message indicating the path of the file that was created.

The PDF or PNG is created via \code{\link[grDevices:pdf]{grDevices::pdf()}} or \code{\link[grDevices:png]{grDevices::png()}}, respectively.
Fonts and lines are bolded by setting all of the following parameters to \code{2} via \code{\link[graphics:par]{graphics::par()}}:
\code{font}, \code{font.main}, \code{font.sub}, \code{font.lab}, \code{font.axis},  and \code{lwd}.
Similarly, titles and axis labels are placed closer to the figure by setting \code{mgp = c(2, 0.5, 0)},
and the approximate number of tickmarks is reduced by setting \code{lab = c(3, 3, 7)}.}
}
\description{
Creates a PDF or PNG figure with dimensions that default approximately to a single column on a two-column paper, bolds text, reduces space between labels and the figure, and sets other improved defaults for margins.
The goal is to create publication- and presentation-quality figures with few keystrokes.
PNG version is by default high resolution (600 ppi) and scales dimensions according to this resolution to get images that match their PDF counterparts as much as possible.
}
\examples{
### simple square figure with nice margins
# fig_start('my_plot')
# plot(1:10)
# fig_end()

### example with a title
### increase top margin only
# fig_start('my_plot', mar_t = 1.5)
# plot(1:10, main = 'My Title!')
# fig_end()

}
